Refactor metadata helpers into focused modules
